chore: Registry tests cleanup (#4232)
