+++ sdk/python/tests/integration/registration/test_feature_store.py
@@ -11,66 +11,19 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import os
-import time
 from datetime import timedelta
 from tempfile import mkstemp
 
 import pytest
 from pytest_lazyfixture import lazy_fixture
 
-from feast import FileSource
-from feast.data_format import ParquetFormat
 from feast.entity import Entity
 from feast.feature_store import FeatureStore
 from feast.feature_view import FeatureView
-from feast.field import Field
-from feast.infra.offline_stores.file import FileOfflineStoreConfig
-from feast.infra.online_stores.dynamodb import DynamoDBOnlineStoreConfig
 from feast.infra.online_stores.sqlite import SqliteOnlineStoreConfig
 from feast.repo_config import RepoConfig
-from feast.types import Array, Bytes, Float64, Int64, String
+from feast.types import Float64, Int64, String
 from tests.utils.data_source_test_creator import prep_file_source
-
-
-@pytest.mark.integration
-@pytest.mark.parametrize(
-    "test_feature_store",
-    [
-        lazy_fixture("feature_store_with_gcs_registry"),
-        lazy_fixture("feature_store_with_s3_registry"),
-    ],
-)
-def test_apply_entity_integration(test_feature_store):
-    entity = Entity(
-        name="driver_car_id",
-        description="Car driver id",
-        tags={"team": "matchmaking"},
-    )
-
-    # Register Entity
-    test_feature_store.apply([entity])
-
-    entities = test_feature_store.list_entities()
-
-    entity = entities[0]
-    assert (
-        len(entities) == 1
-        and entity.name == "driver_car_id"
-        and entity.description == "Car driver id"
-        and "team" in entity.tags
-        and entity.tags["team"] == "matchmaking"
-    )
-
-    entity = test_feature_store.get_entity("driver_car_id")
-    assert (
-        entity.name == "driver_car_id"
-        and entity.description == "Car driver id"
-        and "team" in entity.tags
-        and entity.tags["team"] == "matchmaking"
-    )
-
-    test_feature_store.teardown()
 
 
 @pytest.mark.integration
@@ -109,81 +62,6 @@ def test_feature_view_inference_success(test_feature_store, dataframe_source):
         test_feature_store.teardown()
 
 
-@pytest.mark.integration
-@pytest.mark.parametrize(
-    "test_feature_store",
-    [
-        lazy_fixture("feature_store_with_gcs_registry"),
-        lazy_fixture("feature_store_with_s3_registry"),
-    ],
-)
-def test_apply_feature_view_integration(test_feature_store):
-    # Create Feature Views
-    batch_source = FileSource(
-        file_format=ParquetFormat(),
-        path="file://feast/*",
-        timestamp_field="ts_col",
-        created_timestamp_column="timestamp",
-    )
-
-    entity = Entity(name="fs1_my_entity_1", join_keys=["test"])
-
-    fv1 = FeatureView(
-        name="my_feature_view_1",
-        schema=[
-            Field(name="fs1_my_feature_1", dtype=Int64),
-            Field(name="fs1_my_feature_2", dtype=String),
-            Field(name="fs1_my_feature_3", dtype=Array(String)),
-            Field(name="fs1_my_feature_4", dtype=Array(Bytes)),
-            Field(name="test", dtype=Int64),
-        ],
-        entities=[entity],
-        tags={"team": "matchmaking"},
-        source=batch_source,
-        ttl=timedelta(minutes=5),
-    )
-
-    # Register Feature View
-    test_feature_store.apply([fv1, entity])
-
-    feature_views = test_feature_store.list_feature_views()
-
-    # List Feature Views
-    assert (
-        len(feature_views) == 1
-        and feature_views[0].name == "my_feature_view_1"
-        and feature_views[0].features[0].name == "fs1_my_feature_1"
-        and feature_views[0].features[0].dtype == Int64
-        and feature_views[0].features[1].name == "fs1_my_feature_2"
-        and feature_views[0].features[1].dtype == String
-        and feature_views[0].features[2].name == "fs1_my_feature_3"
-        and feature_views[0].features[2].dtype == Array(String)
-        and feature_views[0].features[3].name == "fs1_my_feature_4"
-        and feature_views[0].features[3].dtype == Array(Bytes)
-        and feature_views[0].entities[0] == "fs1_my_entity_1"
-    )
-
-    feature_view = test_feature_store.get_feature_view("my_feature_view_1")
-    assert (
-        feature_view.name == "my_feature_view_1"
-        and feature_view.features[0].name == "fs1_my_feature_1"
-        and feature_view.features[0].dtype == Int64
-        and feature_view.features[1].name == "fs1_my_feature_2"
-        and feature_view.features[1].dtype == String
-        and feature_view.features[2].name == "fs1_my_feature_3"
-        and feature_view.features[2].dtype == Array(String)
-        and feature_view.features[3].name == "fs1_my_feature_4"
-        and feature_view.features[3].dtype == Array(Bytes)
-        and feature_view.entities[0] == "fs1_my_entity_1"
-    )
-
-    test_feature_store.delete_feature_view("my_feature_view_1")
-    feature_views = test_feature_store.list_feature_views()
-    assert len(feature_views) == 0
-
-    test_feature_store.teardown()
-
-
 @pytest.fixture
 def feature_store_with_local_registry():
     fd, registry_path = mkstemp()
@@ -194,49 +72,6 @@ def feature_store_with_local_registry():
             project="default",
             provider="local",
             online_store=SqliteOnlineStoreConfig(path=online_store_path),
-            entity_key_serialization_version=2,
-        )
-    )
-
-
-@pytest.fixture
-def feature_store_with_gcs_registry():
-    from google.cloud import storage
-
-    storage_client = storage.Client()
-    bucket_name = f"feast-registry-test-{int(time.time() * 1000)}"
-    bucket = storage_client.bucket(bucket_name)
-    bucket = storage_client.create_bucket(bucket)
-    bucket.add_lifecycle_delete_rule(
-        age=14
-    )  # delete buckets automatically after 14 days
-    bucket.patch()
-    bucket.blob("registry.db")
-
-    return FeatureStore(
-        config=RepoConfig(
-            registry=f"gs://{bucket_name}/registry.db",
-            project="default",
-            provider="gcp",
-            entity_key_serialization_version=2,
-        )
-    )
-
-
-@pytest.fixture
-def feature_store_with_s3_registry():
-    aws_registry_path = os.getenv(
-        "AWS_REGISTRY_PATH", "s3://feast-int-bucket/registries"
-    )
-    return FeatureStore(
-        config=RepoConfig(
-            registry=f"{aws_registry_path}/{int(time.time() * 1000)}/registry.db",
-            project="default",
-            provider="aws",
-            online_store=DynamoDBOnlineStoreConfig(
-                region=os.getenv("AWS_REGION", "us-west-2")
-            ),
-            offline_store=FileOfflineStoreConfig(),
             entity_key_serialization_version=2,
         )
     )
